add magpie bsc, small refacto, split pools
